Create private_objects testing directory
