+++ python/cudf/cudf/tests/private_objects/test_column.py
@@ -5,31 +5,17 @@ import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
+from numba import cuda
 
 import cudf
 from cudf.core.column.column import _can_values_be_equal, as_column
 from cudf.testing import assert_eq
 from cudf.testing._utils import assert_exceptions_equal
-from cudf.utils import dtypes as dtypeutils
-
-dtypes = sorted(
-    list(
-        dtypeutils.ALL_TYPES
-        - {
-            "datetime64[s]",
-            "datetime64[ms]",
-            "datetime64[us]",
-            "timedelta64[s]",
-            "timedelta64[ms]",
-            "timedelta64[us]",
-        }
-    )
-)
 
 
-@pytest.fixture(params=dtypes, ids=dtypes)
-def pandas_input(request):
-    dtype = request.param
+@pytest.fixture
+def pandas_input(all_supported_types_as_str):
+    dtype = all_supported_types_as_str
     rng = np.random.default_rng(seed=0)
     size = 100
 
@@ -72,7 +58,7 @@ def str_host_view(list_of_str, to_dtype):
 @pytest.mark.parametrize("offset", [0, 1, 15])
 @pytest.mark.parametrize("size", [50, 10, 0])
 def test_column_offset_and_size(pandas_input, offset, size):
-    col = cudf.core.column.as_column(pandas_input)
+    col = as_column(pandas_input)
     col = cudf.core.column.build_column(
         data=col.base_data,
         dtype=col.dtype,
@@ -141,7 +127,7 @@ def column_slicing_test(col, offset, size, cast_to_float=False):
 @pytest.mark.parametrize("offset", [0, 1, 15])
 @pytest.mark.parametrize("size", [50, 10, 0])
 def test_column_slicing(pandas_input, offset, size):
-    col = cudf.core.column.as_column(pandas_input)
+    col = as_column(pandas_input)
     column_slicing_test(col, offset, size)
 
 
@@ -154,9 +140,7 @@ def test_column_slicing(pandas_input, offset, size):
     [cudf.Decimal128Dtype, cudf.Decimal64Dtype, cudf.Decimal32Dtype],
 )
 def test_decimal_column_slicing(offset, size, precision, scale, decimal_type):
-    col = cudf.core.column.as_column(
-        pd.Series(np.random.default_rng(seed=0).random(1000))
-    )
+    col = as_column(pd.Series(np.random.default_rng(seed=0).random(1000)))
     col = col.astype(decimal_type(precision, scale))
     column_slicing_test(col, offset, size, True)
 
@@ -173,7 +157,7 @@ def test_column_series_multi_dim(data):
         cudf.Series(data)
 
     with pytest.raises(ValueError):
-        cudf.core.column.as_column(data)
+        as_column(data)
 
 
 @pytest.mark.parametrize(
@@ -195,14 +179,13 @@ def test_column_mixed_dtype(data, error):
             cudf.Series(data)
 
 
-@pytest.mark.parametrize("nan_as_null", [True, False])
 @pytest.mark.parametrize(
     "scalar",
     [np.nan, pd.Timedelta(days=1), pd.Timestamp(2020, 1, 1)],
     ids=repr,
 )
-@pytest.mark.parametrize("size", [1, 10])
-def test_as_column_scalar_with_nan(nan_as_null, scalar, size):
+def test_as_column_scalar_with_nan(nan_as_null, scalar):
+    size = 5
     expected = (
         cudf.Series([scalar] * size, nan_as_null=nan_as_null)
         .dropna()
@@ -221,23 +204,20 @@ def test_as_column_scalar_with_nan(nan_as_null, scalar, size):
 
 
 @pytest.mark.parametrize("data", [[1.1, 2.2, 3.3, 4.4], [1, 2, 3, 4]])
-@pytest.mark.parametrize("dtype", ["float32", "float64"])
-def test_column_series_cuda_array_dtype(data, dtype):
-    psr = pd.Series(np.asarray(data), dtype=dtype)
-    sr = cudf.Series(cp.asarray(data), dtype=dtype)
+def test_column_series_cuda_array_dtype(data, float_types_as_str):
+    psr = pd.Series(np.asarray(data, dtype=float_types_as_str))
+    sr = cudf.Series(cp.asarray(data, dtype=float_types_as_str))
 
     assert_eq(psr, sr)
 
-    psr = pd.Series(data, dtype=dtype)
-    sr = cudf.Series(data, dtype=dtype)
+    psr = pd.Series(data, dtype=float_types_as_str)
+    sr = cudf.Series(data, dtype=float_types_as_str)
 
     assert_eq(psr, sr)
 
 
 def test_column_zero_length_slice():
     # see https://github.com/rapidsai/cudf/pull/4777
-    from numba import cuda
-
     x = cudf.DataFrame({"a": [1]})
     the_column = x[1:]["a"]._column
 
@@ -251,20 +231,16 @@ def test_column_chunked_array_creation():
     pyarrow_array = pa.array([1, 2, 3] * 1000)
     chunked_array = pa.chunked_array(pyarrow_array)
 
-    actual_column = cudf.core.column.as_column(
-        chunked_array, dtype=np.dtype(np.float64)
-    )
-    expected_column = cudf.core.column.as_column(
-        pyarrow_array, dtype=np.dtype(np.float64)
-    )
+    actual_column = as_column(chunked_array, dtype=np.dtype(np.float64))
+    expected_column = as_column(pyarrow_array, dtype=np.dtype(np.float64))
 
     assert_eq(
         cudf.Series._from_column(actual_column),
         cudf.Series._from_column(expected_column),
     )
 
-    actual_column = cudf.core.column.as_column(chunked_array)
-    expected_column = cudf.core.column.as_column(pyarrow_array)
+    actual_column = as_column(chunked_array)
+    expected_column = as_column(pyarrow_array)
 
     assert_eq(
         cudf.Series._from_column(actual_column),
@@ -418,8 +394,8 @@ def test_column_view_string_slice(slc):
     ],
 )
 def test_as_column_buffer(box, data):
-    expected = cudf.core.column.as_column(data)
-    actual_column = cudf.core.column.as_column(
+    expected = as_column(data)
+    actual_column = as_column(
         cudf.core.buffer.as_buffer(box(data)), dtype=data.dtype
     )
     assert_eq(
@@ -528,29 +504,6 @@ def test_build_series_from_nullable_pandas_dtype(pd_dtype, expect_dtype):
     got_mask = gd_data._column._get_mask_as_column().values_host
 
     np.testing.assert_array_equal(expect_mask, got_mask)
-
-
-@pytest.mark.parametrize(
-    "alias,expect_dtype",
-    [
-        ("UInt8", "uint8"),
-        ("UInt16", "uint16"),
-        ("UInt32", "uint32"),
-        ("UInt64", "uint64"),
-        ("Int8", "int8"),
-        ("Int16", "int16"),
-        ("Int32", "int32"),
-        ("Int64", "int64"),
-        ("boolean", "bool"),
-        ("Float32", "float32"),
-        ("Float64", "float64"),
-    ],
-)
-def test_astype_with_aliases(alias, expect_dtype):
-    pd_data = pd.Series([1, 2, 0])
-    gd_data = cudf.Series.from_pandas(pd_data)
-
-    assert_eq(pd_data.astype(expect_dtype), gd_data.astype(alias))
 
 
 @pytest.mark.parametrize(
