+++ python/cudf/cudf/tests/series/methods/test_astype.py
@@ -163,6 +163,29 @@ def test_numeric_to_timedelta(
     assert_eq(expected, actual)
 
 
+@pytest.mark.parametrize(
+    "alias,expect_dtype",
+    [
+        ("UInt8", "uint8"),
+        ("UInt16", "uint16"),
+        ("UInt32", "uint32"),
+        ("UInt64", "uint64"),
+        ("Int8", "int8"),
+        ("Int16", "int16"),
+        ("Int32", "int32"),
+        ("Int64", "int64"),
+        ("boolean", "bool"),
+        ("Float32", "float32"),
+        ("Float64", "float64"),
+    ],
+)
+def test_astype_with_aliases(alias, expect_dtype):
+    pd_data = pd.Series([1, 2, 0])
+    gd_data = cudf.Series.from_pandas(pd_data)
+
+    assert_eq(pd_data.astype(expect_dtype), gd_data.astype(alias))
+
+
 def test_timedelta_datetime_cast_invalid():
     sr = cudf.Series([1, 2, 3], dtype="timedelta64[ns]")
     psr = sr.to_pandas()
